+++ kw-uni/History/HistoryStateBase.cpp
@@ -119,13 +119,13 @@ namespace {
             } else if (prevOut.empty()) {
                 // ②検索が実行されたが、出力文字列にはキーだけが表示されている状態
                 _LOG_DEBUGH(_T("CURRENT: SetOutString(str={}, numBS={})"), to_wstr(prevKey), prevKey.size());
-                resultStr.setResult(prevKey, prevKey.size());
+                resultStr.setResult(prevKey, (int)(prevKey.size()));
                 STROKE_MERGER_NODE->SetPrevHistState(prevKey, prevKey);
                 _LOG_DEBUGH(_T("CURRENT: prevKey={}"), to_wstr(prevKey));
             } else {
                 // ③横列のどれかの候補が選択されて出力文字列に反映されている状態
                 _LOG_DEBUGH(_T("REVERT and NEW HIST: SetOutString(str={}, numBS={})"), to_wstr(prevKey), prevOut.size());
-                resultStr.setResult(prevKey, prevOut.size());
+                resultStr.setResult(prevKey, (int)(prevOut.size()));
                 STROKE_MERGER_NODE->SetPrevHistState(prevKey, prevKey);
                 _LOG_DEBUGH(_T("REVERT and NEW HIST: prevKey={}"), to_wstr(prevKey));
             }
